chore(windows): move keymanversion.h to common
