+++ common/windows/cpp/include/keymanversion.h
