+++ common/windows/cpp/include/keymanversion_build.in
