+++ windows/src/Target.mak
@@ -21,11 +21,11 @@ global-versions:
 # Build our root version.txt
     cd $(ROOT)\src
 
-    $(MKVER_U) $(ROOT)\src\global\inc\keymanversion_build.in $(ROOT)\src\global\inc\keymanversion_build.h
+    $(MKVER_U)$(KEYMAN_ROOT)\common\windows\cpp\include\keymanversion_build.in $(KEYMAN_ROOT)\common\windows\cpp\include\keymanversion_build.h
     $(MKVER_U) $(ROOT)\src\global\delphi\general\keymanversion_build.in $(ROOT)\src\global\delphi\general\keymanversion_build.inc
 
 devtools-app: dirs
-    @if not exist $(WINDOWS_ROOT)\src\global\inc\keymanversion_build.h $(MAKE) global-versions
+    @if not exist $(KEYMAN_ROOT)\common\windows\cpp\include\keymanversion_build.h $(MAKE) global-versions
     cd $(COMMON_ROOT)\tools\devtools
     $(MAKE)
     cd $(ROOT)\src
