+++ windows/src/desktop/kmconfig/version.rc
@@ -1,4 +1,4 @@
-#include "../../global/inc/keymanversion.h"
+#include "../../../../common/windows/cpp/include/keymanversion.h"
 
 1 VERSIONINFO
   FILEVERSION		KV_FILEVERSION
