+++ windows/src/global/Makefile
@@ -17,7 +17,7 @@ build-release:
 
 clean:
   -del $(ROOT)\src\global\delphi\general\keymanversion_build.inc
-  -del $(ROOT)\src\global\inc\keymanversion_build.h
+  -del $(KEYMAN_ROOT)\common\windows\cpp\include\keymanversion_build.h
 
   cd $(ROOT)\src\global\delphi
   $(MAKE) clean
